# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Family Greenhouse API
  version: 1.0.0
  description: |
    Read-only access to your household's plants, tasks, and activity feed.
    Designed for personal automations (Home Assistant, scripts, dashboards).
    Write access is not provided in v1 — use the web app to make changes.
servers:
  - url: https://api.family-greenhouse.example/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Local dev server
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: fg_
  schemas:
    Plant:
      type: object
      required: [id, householdId, name, createdAt, createdBy, updatedAt]
      properties:
        id: { type: string, format: uuid }
        householdId: { type: string, format: uuid }
        name: { type: string, maxLength: 100 }
        species: { type: string, nullable: true, maxLength: 100 }
        location: { type: string, nullable: true, maxLength: 100 }
        imageUrl: { type: string, nullable: true, format: uri }
        notes: { type: string, nullable: true, maxLength: 1000 }
        tags:
          type: array
          items: { type: string, maxLength: 40 }
          maxItems: 10
        createdAt: { type: string, format: date-time }
        createdBy: { type: string }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
security:
  - apiKey: []
paths:
...
  /plants/{id}:
    get:
      summary: Get a single plant by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Plant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plant' }
        '404':
          description: Plant not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...